requirements.txt
numpy
pandas
matplotlib
pyyaml
sklearn
pytorch
# For mac
tensorflow-macos
tensorflow-metal
